Move RouteService methods to MapService
